Store the block associated with the devices (device + attached lcd)
Parse $ style variables from the custom name of the lcd to influence behaviour.
Right now just using fnt for font size (which is kind of not very useful but just for testing)

Some additional spew to catch a bad cleanup case.

NukeLCD non static as it will now clean up positions and blocks as well when a device or lcd is destroyed or picked up with a multitool